fix: split oversized runtime child test helpers
