+++ pkg/orchestrators/javascript/runtime/host_children_test.go
@@ -965,84 +965,3 @@ func TestRun_ParallelFakeChildren_RepresentsFailedChildExplicitly(t *testing.T) 
 		t.Fatal("failed child should not emit COMPLETED child_dispatch record")
 	}
 }
-
-func assertParallelResultOrder(t *testing.T, sessionID string, outcome workflowruntime.Outcome, wantLabels []string) {
-	t.Helper()
-	projected := projectPrimaryJSON(t, sessionID, outcome.Value)
-	results, ok := projected["results"].([]any)
-	if !ok || len(results) != len(wantLabels) {
-		t.Fatalf("projected results = %#v, want %d entries", projected["results"], len(wantLabels))
-	}
-	for i, wantLabel := range wantLabels {
-		child, ok := results[i].(map[string]any)
-		if !ok {
-			t.Fatalf("results[%d] = %#v, want object", i, results[i])
-		}
-		if child["label"] != wantLabel {
-			t.Fatalf("results[%d].label = %#v, want %q", i, child["label"], wantLabel)
-		}
-		if child["status"] != workflowruntime.ChildDispatchStatusCompleted {
-			t.Fatalf("results[%d].status = %#v, want COMPLETED", i, child["status"])
-		}
-	}
-}
-
-func peakConcurrentChildRunning(records []workflowruntime.RuntimeRecord) int {
-	running := 0
-	peak := 0
-	for _, record := range records {
-		if record.Kind != workflowruntime.RecordKindChildDispatch || record.ChildDispatch == nil {
-			continue
-		}
-		switch record.ChildDispatch.Status {
-		case workflowruntime.ChildDispatchStatusRunning:
-			running++
-			if running > peak {
-				peak = running
-			}
-		case workflowruntime.ChildDispatchStatusCompleted, workflowruntime.ChildDispatchStatusFailed:
-			running--
-		}
-	}
-	return peak
-}
-
-func childCompletionOrder(records []workflowruntime.RuntimeRecord) []string {
-	var order []string
-	for _, record := range records {
-		if record.Kind != workflowruntime.RecordKindChildDispatch || record.ChildDispatch == nil {
-			continue
-		}
-		if record.ChildDispatch.Status != workflowruntime.ChildDispatchStatusCompleted {
-			continue
-		}
-		if record.ChildDispatch.Label != "" {
-			order = append(order, record.ChildDispatch.Label)
-		}
-	}
-	return order
-}
-
-func hasChildDispatchStatus(records []workflowruntime.RuntimeRecord, status string) bool {
-	for _, record := range records {
-		if record.Kind == workflowruntime.RecordKindChildDispatch &&
-			record.ChildDispatch != nil &&
-			record.ChildDispatch.Status == status {
-			return true
-		}
-	}
-	return false
-}
-
-func completedForLabel(records []workflowruntime.RuntimeRecord, label string) bool {
-	for _, record := range records {
-		if record.Kind != workflowruntime.RecordKindChildDispatch || record.ChildDispatch == nil {
-			continue
-		}
-		if record.ChildDispatch.Label == label &&
-			record.ChildDispatch.Status == workflowruntime.ChildDispatchStatusCompleted {
-			return true
-		}
-	}
-	return false
-}
